// pdlearning/src/web-apps/src/libraries/domain-components/src/models/course-tab.model.ts
export enum CourseTabInfo {
  OverviewInfo = 't-overview-info',
  ProviderInfo = 't-provider-info',
  MetaData = 't-metadata',
  MetaDataCopyRight = 't-copyright',
  TargetAudience = 't-target-audience',
  CoursePlanning = 't-course-planning',
  CourseAdministration = 't-course-administration',
  EvaluationEcertificate = 't-evaluation-E-Certificate'
}